Few more Jitc intructions

+++ SRC/Hardware/CP.h
@@ -146,7 +146,7 @@ typedef struct FifoControl
 {
     CPRegs      cp;     // command processor registers
     PERegs      pe;     // pixel engine registers
-    uint32_t    done_num;   // number of drawdone (PE_FINISH) events
+    size_t      done_num;   // number of drawdone (PE_FINISH) events
     bool        log;
     Thread*     thread;     // CP FIFO thread
     size_t      tickPerFifo;
